# PERF: speed up nanmin/max with non-null axis for int32